rename some utilities to avoid confusion with built-in functions
